refactor(core): split blueprint.rs by concept
1,571 production lines holding four things a reader looks for separately: what
an agent is, what a stage is, which model it runs on, and how it leaves one
stage for the next.

- `blueprint/stage.rs` (663) - `Stage` and `StageMode`, with each mode's config
  beside the mode it configures rather than in a flat bag on `Stage`. Modes are
  what make stages differ in kind: autonomous loops on tools, fan-out splits
  into workers, interaction-points stops for a person.
- `blueprint/transition.rs` (348) - edges, conditions, transforms, gates, nudges
  and stuck detection. Nudges and stuck detection belong here because they exist
  to answer the same question an edge does: this stage is not finishing, now
  what.
- `blueprint/model.rs` (82) - `ModelEntry` and `ModelConfig`.
- `blueprint/mod.rs` - `Blueprint` itself, `SafeCommandsConfig`,
  `ReadPathsConfig`, and validation.

Glob re-exported per `pipeline/mod.rs`, so every existing `blueprint::Stage`
path is unchanged and no call site moved. The submodules reach each other
through the parent's re-exports (`use super::*`), so a type moving between
sections later does not touch an import list.

**Verified pure by sorted-line diff: 1,940 in, 1,940 out.** The entire
difference is `Stage::validate` and `ContextTransform::validate` gaining
`pub(super)`, which the move forces: `Blueprint::validate` calls both and they
now live one module over.

Worth recording, since the config split hit the same thing an hour earlier: a
**stale `target/llvm-cov-target` reports uncovered lines in a file that no longer
exists**. It named `blueprint.rs:179` after the split. `rm -rf` it after any
rename before believing a coverage result.

771 core tests pass and the package holds 100%.
